refactor(api): refactor directory structure
