+++ apps/api/src/domain/models/vocabulary-list/entities/vocabulary-list.entity.ts
@@ -1,8 +1,8 @@
 import isStringWithNonzeroLength from 'apps/api/src/lib/utilities/isStringWithNonzeroLength';
 import { PartialDTO } from 'apps/api/src/types/partial-dto';
-import { Entity } from '../../models/entity';
-import { entityTypes } from '../../types/entityType';
-import { determineAllMissingRequiredProperties } from '../../utilities/validation/determine-all-missing-required-properties';
+import { entityTypes } from '../../../types/entityType';
+import { determineAllMissingRequiredProperties } from '../../../utilities/validation/determine-all-missing-required-properties';
+import { Entity } from '../../entity';
 import { VocabularyListVariable } from '../types/vocabulary-list-variable';
 import { VocabularyListEntry } from '../vocabulary-list-entry';
