+++ apps/api/src/domain/models/vocabulary-list/types/dropbox-or-checkbox.ts
