+++ apps/api/src/domain/models/vocabulary-list/types/vocabulary-list-variable-value.ts
