+++ apps/api/src/domain/models/vocabulary-list/vocabulary-list-entry.ts
@@ -1,4 +1,4 @@
-import { EntityId } from '../types/entity-id';
+import { EntityId } from '../../types/entity-id';
 import { VocabularyListVariableValue } from './types/vocabulary-list-variable-value';
 
 export class VocabularyListEntry {
