+++ apps/api/src/domain/types/all-entities.ts
@@ -1,6 +1,6 @@
 import { PartialDTO } from '../../types/partial-dto';
 import { Term } from '../models/term/entities/term.entity';
-import { VocabularyList } from '../vocabulary-list/entities/vocabulary-list.entity';
+import { VocabularyList } from '../models/vocabulary-list/entities/vocabulary-list.entity';
 
 /**
  * TODO [refactor] Remove this in favor of
