+++ apps/api/src/view-models/buildViewModelForEntity/viewModels/VocabularyListViewModel.ts
@@ -1,8 +1,8 @@
 import { Term } from '../../../domain/models/term/entities/term.entity';
+import { VocabularyList } from '../../../domain/models/vocabulary-list/entities/vocabulary-list.entity';
+import { VocabularyListVariable } from '../../../domain/models/vocabulary-list/types/vocabulary-list-variable';
+import { VocabularyListVariableValue } from '../../../domain/models/vocabulary-list/types/vocabulary-list-variable-value';
 import { EntityId } from '../../../domain/types/entity-id';
-import { VocabularyList } from '../../../domain/vocabulary-list/entities/vocabulary-list.entity';
-import { VocabularyListVariable } from '../../../domain/vocabulary-list/types/vocabulary-list-variable';
-import { VocabularyListVariableValue } from '../../../domain/vocabulary-list/types/vocabulary-list-variable-value';
 import { NotFound } from '../../../lib/types/not-found';
 import { TermViewModel } from './TermViewModel';
